+++ packages/ragbits-core/CHANGELOG.md
@@ -1,7 +1,13 @@
 # CHANGELOG
 
 ## Unreleased
+
+## 0.15.0 (2025-04-28)
 - Allow using sparse embeddings with Qdrant and local vector stores (#493)
+- Add support for sparse embeddings in the Pgvector Vector Store (#493)
+- Added secret key environment variable / generation for signatures across ragbits packages (#537)
+
+- Fix source interface definition (#535)
 - Configured source and element classes are auto-imported (#343)
 
 ## 0.14.0 (2025-04-22)